ENH: fixing the default library lookups in the site.cfg
Looking for the libraries in system_info is still very difficult,
this patch, tries to work-around some of the more problematic issues
by stream-lining the library look-ups in the site.cfg.
E.g. previously

[mkl]
libraries = foo
mkl_libs = mkl_rt

it would only search for mkl_libs.
Now it will first search for libraries key, then use that as a default
for the mkl_libs key. Generally, I think the mkl_libs key is superfluous,
and should be removed completely (there is no ambiguity in the libraries
key which is under the [mkl] section. It *must* be a MKL library!).

This methodology has been fixed in a number of libraries:
  mkl
  openblas
  blas
  fftw[23]
  blis
  amd
  umfpack

Signed-off-by: Nick Papior <nickpapior@gmail.com>